refactor code to avoid circular import and fix base command exception
